+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the create-package project will be documented in this fil
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- Latest package version numbers in package.json will be automatically fetched.
+
 ## [2.3.0] - 2019-03-27
 ### Added
 - More initial setup explanation in README.md.